# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Personal Finance Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions with optional filters
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: merchant
          in: query
          schema:
            type: string
        - name: needsReview
          in: query
          schema:
            type: boolean
        - name: type
          in: query
          schema:
            type: string
        - name: accountId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [date, amount, category, merchant]
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
...
  /transactions/bulk-categorize:
    post:
      operationId: bulkCategorize
      tags: [transactions]
      summary: Update the category for multiple transactions at once
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkCategorizeRequest"
      responses:
        "200":
          description: Number of updated transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkCategorizeResult"
  /transactions/bulk-review:
    post:
      operationId: bulkReviewTransactions
      tags: [transactions]
      summary: Mark multiple transactions as reviewed or needing review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkReviewTransactionsRequest"
      responses:
        "200":
          description: Number of updated transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkReviewTransactionsResult"
  /transactions/{id}:
    patch:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction category or notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [error]
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        merchant:
          type: string
        description:
          type: string
        amount:
          type: number
        accountId:
          type: ["integer", "null"]
        accountName:
          type: ["string", "null"]
        accountType:
          type: ["string", "null"]
        type:
          type: string
          enum: [income, expense, transfer, refund, debit, credit]
        direction:
          type: string
          enum: [debit, credit]
        currency:
          type: string
        transactionKind:
          type: string
        transferGroupId:
          type: ["string", "null"]
        matchedTransferId:
          type: ["integer", "null"]
        bank:
          type: string
        parser:
          type: ["string", "null"]
        balance:
          type: ["number", "null"]
        category:
          type: string
        categorizationConfidence:
          type: number
        categorizationSource:
          type: string
        categorizationExplanation:
          type: string
        importConfidence:
          type: number
        month:
          type: string
        notes:
          type: ["string", "null"]
        reviewed:
          type: boolean
        createdAt:
          type: string
      required: [id, date, merchant, description, amount, type, direction, currency, transactionKind, bank, balance, category, categorizationConfidence, categorizationSource, categorizationExplanation, importConfidence, month, reviewed, createdAt]
    TransactionList:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
      required: [transactions, total]
    TransactionUpdate:
      type: object
      properties:
        category:
          type: string
        notes:
          type: string
        reviewed:
          type: boolean
    BulkCategorizeRequest:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
        category:
          type: string
      required: [ids, category]
    BulkCategorizeResult:
      type: object
      properties:
        updated:
          type: integer
      required: [updated]
    BulkReviewTransactionsRequest:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
        reviewed:
          type: boolean
      required: [ids, reviewed]
    BulkReviewTransactionsResult:
      type: object
      properties:
        updated:
          type: integer
      required: [updated]
...